SignalHandling class framework implemented.
The interface and a little of the implementation for the SignalHandling class in windwos is added for this change.  The main implementation detail included is which signals are supported by mingw.  Otherwise, the implementation is equivalent to the Linux one, but without the unsupported data structures; stack_t and sigaction are POSIX and not supported on Windows.